Tests refactoring
