+++ crates/tako/src/internal/worker/resources/mod.rs
@@ -2,3 +2,6 @@ pub mod allocator;
 pub mod concise;
 pub mod map;
 pub mod pool;
+
+#[cfg(test)]
+mod test_allocator;
